Add mutable robin hood hashmap types for holding descriptors (cljs only)
This is purely to try to get more speed out of the browser by using
mutation. Both act like transients except they are not persistable,
hashable, or seqable. However, they are iterable and thus reducible,
and the map implements reduce-kv also.

DescriptorMap is suitable for holding a graph-value, although it should
be wrapped in an ilookup-only proxy for tracker use.

DescriptorSet uses a DescriptorMap; it may be suitable for maintaining
descriptor parent/child sets.